Generischer. Noch generischer. AM generischsten...

+++ sessionOpt/src/main/java/sessionOpt/entities/Prerequisite.java
@@ -2,6 +2,6 @@ package sessionOpt.entities;
 
 public interface Prerequisite {
 	public int getUnsatisfiedPenalty();
-	public int getHappiness(Feature feature);
+	public double getHappiness(Feature feature);
 	public String getName();
 }
